src/models/svd_collaborative.py
"""
Layer 2: Collaborative Filtering with SVD (Singular Value Decomposition).
Uses Surprise library for matrix factorization to discover latent factors.
Re-ranks candidates from Layer 1 based on predicted ratings.
"""

from surprise import SVD, Dataset, Reader


class CollaborativeFilteringSVD:
    """
    SVD-based collaborative filtering using Surprise library.
    Learns latent factors from user-item rating matrix.
    """
    
    def __init__(self, n_factors=50, n_epochs=20, lr=0.005, reg=0.02):
        """
        Initialize SVD model with hyperparameters.
        
        Args:
            n_factors: Number of latent factors (20-100)
            n_epochs: Number of training epochs (10-30)
            lr: Learning rate for SGD
            reg: Regularization strength
        """
        pass
    
    def train(self, train_data):
        """
        Train SVD model on training data.
        
        Args:
            train_data: DataFrame with columns [user_id, item_id, rating]
        """
        pass
    
    def predict(self, user_id, item_id):
        """
        Predict rating for a user-item pair.
        
        Returns:
            float: Predicted rating (1-5 scale)
        """
        pass
    
    def recommend(self, user_id, candidate_items, top_k=50):
        """
        Re-rank candidate items from Layer 1 by predicted ratings.
        
        Args:
            user_id: User ID
            candidate_items: List of movie IDs from content-based layer
            top_k: Number of top items to return
        
        Returns:
            list: Top-K movie IDs sorted by predicted rating
        """
        pass

