src/visualization/embeddings.py
"""
Embedding visualization using dimensionality reduction.
Visualizes high-dimensional movie/user embeddings in 2D/3D space.
Supports t-SNE and UMAP for non-linear projection.
"""

import numpy as np
import matplotlib.pyplot as plt
from sklearn.manifold import TSNE


def visualize_embeddings(embeddings, labels=None, method='tsne', 
                         n_components=2, title='Embeddings'):
    """
    Visualize high-dimensional embeddings in 2D/3D.
    
    Args:
        embeddings: numpy array of shape (n_samples, embedding_dim)
        labels: Optional labels for coloring points
        method: 'tsne' or 'umap'
        n_components: 2 or 3 for 2D/3D visualization
        title: Plot title
    
    Returns:
        matplotlib.figure.Figure
    """
    pass


def plot_embedding_evolution(embeddings_history, labels):
    """
    Show how embeddings change during training.
    Creates animation of embedding space over epochs.
    """
    pass


def find_similar_items(item_id, embeddings, movie_titles, top_k=10):
    """
    Find most similar items in embedding space.
    Uses cosine similarity or Euclidean distance.
    """
    pass

