requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-surprise>=1.1.3
torch>=2.0.0
scipy>=1.11.0
networkx>=3.1
plotly>=5.14.0
matplotlib>=3.7.0
jupyter>=1.0.0
tqdm>=4.65.0

